feat: centralize landing CTAs and create page
