+++ src/lib/supabase-browser.ts
@@ -1,6 +1,5 @@
 export function getBrowserSupabase() {
   if (typeof window === 'undefined') return null;
-  // singleton on window to avoid multiple clients
   const w = window as any;
   if (w.__sb) return w.__sb;
   const { createClient } = require('@supabase/supabase-js');
@@ -10,3 +9,4 @@ export function getBrowserSupabase() {
   w.__sb = createClient(url, key);
   return w.__sb;
 }
+
